// app/models/calendar.model.ts
export class Calendar {
  id: string;
  name: string;
  cellphone: number;
  date: number;
  monthyear: string;
  time: string;
  description: string;
  expectedvalue: number;
  value: number;
  address: string;
  confirmed: boolean;
  UID: string;
  dateinsert: number;
  email: string;
  service: string;
  status: string;
  infoadd: string;
  calendarid: string;
}